ai-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sse-starlette==2.2.1
langchain==0.3.18
langchain-openai==0.3.6
openai==1.63.2
pymysql==1.1.1
python-dotenv==1.0.1
cryptography==44.0.1
